associate lambda functions to their own files in one /lambdas/ directory, round 1
