+++ .gitignore
@@ -32,11 +32,9 @@ web-api/runtimes/clamav/bin
 web-api/runtimes/clamav/clamav_lambda_layer.tar.gz
 web-api/runtimes/clamav/lib
 web-api/storage/s3/*
-web-api/terraform/template/cognito-authorizer/index.js.zip
+web-api/terraform/template/lambdas/index.js.zip
 web-api/terraform/template/cognito-triggers/index.js.zip
 web-api/terraform/template/log-forwarder/index.js.zip
-web-api/terraform/template/api/index.js.zip
-web-api/terraform/template/api-public/index.js.zip
 web-api/terraform/main/new-policy.json
 web-client/.cache/
 web-client/coverage-e2e
