+++ web-api/terraform/template/cognito-authorizer.tf
@@ -1,14 +1,14 @@
 data "archive_file" "zip_authorizer" {
   type        = "zip"
-  output_path = "${path.module}/lambdas/index.js.zip"
-  source_file = "${path.module}/lambdas/dist/index.js"
+  output_path = "${path.module}/lambdas/cognito-authorizer.js.zip"
+  source_file = "${path.module}/lambdas/dist/cognito-authorizer.js"
 }
 
 resource "aws_lambda_function" "cognito_authorizer_lambda" {
   filename      = "${data.archive_file.zip_authorizer.output_path}"
   function_name = "cognito_authorizer_lambda_${var.environment}"
   role          = "arn:aws:iam::${data.aws_caller_identity.current.account_id}:role/authorizer_lambda_role_${var.environment}"
-  handler       = "index.cognitoAuthorizerHandler"
+  handler       = "cognito-authorizer.handler"
   source_code_hash = "${data.archive_file.zip_authorizer.output_base64sha256}"
   
   runtime = "nodejs12.x"
@@ -21,12 +21,11 @@ resource "aws_lambda_function" "cognito_authorizer_lambda" {
   }
 }
 
-
 resource "aws_lambda_function" "cognito_authorizer_lambda_west" {
   filename      = "${data.archive_file.zip_authorizer.output_path}"
   function_name = "cognito_authorizer_lambda_${var.environment}"
   role          = "arn:aws:iam::${data.aws_caller_identity.current.account_id}:role/authorizer_lambda_role_${var.environment}"
-  handler       = "index.cognitoAuthorizerHandler"
+  handler       = "cognito-authorizer.handler"
   source_code_hash = "${data.archive_file.zip_authorizer.output_base64sha256}"
   
   runtime = "nodejs12.x"
