+++ web-api/terraform/template/cognito-trigger.tf
@@ -1,7 +1,7 @@
 data "archive_file" "zip_triggers" {
   type        = "zip"
-  source_file = "${path.module}/cognito-triggers/dist/index.js"
-  output_path = "${path.module}/cognito-triggers/index.js.zip"
+  source_file = "${path.module}/lambdas/dist/cognito-triggers.js"
+  output_path = "${path.module}/lambdas/cognito-triggers.js.zip"
 }
 
 resource "aws_lambda_permission" "allow_trigger" {
@@ -16,7 +16,7 @@ resource "aws_lambda_function" "cognito_post_confirmation_lambda" {
   filename      = "${data.archive_file.zip_triggers.output_path}"
   function_name = "cognito_post_confirmation_lambda_${var.environment}"
   role          = "arn:aws:iam::${data.aws_caller_identity.current.account_id}:role/iam_cognito_post_confirmation_lambda_role_${var.environment}"
-  handler       = "index.handler"
+  handler       = "cognito-triggers.handler"
   source_code_hash = "${data.archive_file.zip_triggers.output_base64sha256}"
   timeout = "29"
   runtime = "nodejs12.x"
